src/requirements_windows.txt
numpy>=1.26.0
torch
asttokens==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
executing==2.2.0
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.7.0
humanfriendly==10.0
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
netron==8.2.1
networkx==3.5
numpy==2.3.2
onnxruntime-qnn==1.20.1
packaging==25.0
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
protobuf==6.32.0
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.2
pyreadline3==3.5.4
pytest==8.3.5
python-dateutil==2.9.0.post0
pywin32==311
pyzmq==27.0.1
setuptools==75.8.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
torch==2.8.0+cpu
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.14.1
wcwidth==0.2.13
python-dotenv

opencv-python
onnxruntime
onnxruntime-directml

qai-hub
qai-hub-models
# 1) 建立並啟用虛擬環境:
